Se extrae la logica de eliminar un proyecto a un hook aparte
